pyproject.toml
[project]
name = "moodwave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gdeltdoc>=1.12.0",
    "kafka-python>=2.3.0",
    "pyspark>=4.1.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "supabase>=2.24.0",
]

[tool.uv]
package = true
